bugfix on message deletion
Before deleting a message (original or webhook), the client first checks if it is already deleted.

Additional if-block now deletes the nonNitroWebhookMessage if it does not meet the content criteria of a spam channel.

In general, messages are now deleted if the handleNonNitroEmoji handler runs successfully, provided they were not deleted already.

+++ src/events/message.ts
@@ -13,12 +13,20 @@ export default async (client: CommandoClient, message: CommandoMessage) => {
   const nonNitroWebhookMessage = await handleNonNitroEmoji(message);
   const spamEmojiDeleteEmbed = await handleSpamEmojiChannels(message);
   if (spamEmojiDeleteEmbed != null) {
-    message.delete().catch((e) => console.log('Failed to delete a message: ', e));
+    if (message.deleted == false) message.delete().catch((e) => console.log('Failed to delete a message: ', e));
     if (nonNitroWebhookMessage == null) {
       sendDeleteLog(message, spamEmojiDeleteEmbed);
     } else {
       const botSpamEmojiDeleted = await handleSpamEmojiChannels(nonNitroWebhookMessage);
-      if (botSpamEmojiDeleted != null) sendDeleteLog(message, spamEmojiDeleteEmbed);
+      if (botSpamEmojiDeleted != null) {
+        if (nonNitroWebhookMessage.deleted == false) {
+          nonNitroWebhookMessage.delete().catch((e) => console.log('Failed to delete a message: ', e));
+        }
+        sendDeleteLog(message, spamEmojiDeleteEmbed);
+      }
     }
+  }
+  if (nonNitroWebhookMessage != null && message.deleted == false) {
+    message.delete().catch((e) => console.log('Failed to delete a message: ', e));
   }
 };
